feat: finish TS to boards
