+++ src/resources/boards/model.ts
@@ -1,23 +1,24 @@
 import { v4 as uuidv4 } from 'uuid';
 
 import { IBoard, IColumn } from '../../interfaces';
-
-const { generateColumns } = require('./column.model');
+import { generateColumns } from './column.model';
 
 class Board implements IBoard {
   id: string;
 
   title: string;
 
-  columns: IColumn[];
-  /**
-   * @class
-   * @param {string} id
-   * @param {string} title
-   * @param {Array.<Column>} columns
-   */
+  columns: Array<IColumn>;
 
-  constructor({ id = uuidv4(), title = 'BoardTitle', columns = [] } = {}) {
+  constructor({
+    id = uuidv4(),
+    title = 'BoardTitle',
+    columns = [],
+  }: {
+    id: string;
+    title: string;
+    columns: Array<IColumn>;
+  }) {
     this.id = id;
     this.title = title;
     this.columns = generateColumns(columns);
